move test utils to src
