+++ test/background-executor.test.ts
@@ -4,7 +4,7 @@ import { start } from '../src'
 import { Adapter, AdapterEndpoint, EndpointContext } from '../src/adapter'
 import { metrics as eaMetrics } from '../src/metrics'
 import { deferredPromise, sleep } from '../src/util'
-import { NopTransport, NopTransportTypes, TestAdapter } from './util'
+import { NopTransport, NopTransportTypes, TestAdapter } from '../src/util/testing-utils'
 
 const test = untypedTest as TestFn<{
   testAdapter: TestAdapter
