+++ test/error.test.ts
@@ -11,7 +11,7 @@ import {
   AdapterRateLimitError,
   AdapterTimeoutError,
 } from '../src/validation/error'
-import { NopTransport, NopTransportTypes, TestAdapter, assertEqualResponses } from './util'
+import { NopTransport, NopTransportTypes, TestAdapter, assertEqualResponses } from '../src/util/testing-utils'
 
 type TestContext = {
   testAdapter: TestAdapter
