+++ test/subscription-set/subscription-set-factory.test.ts
@@ -4,7 +4,7 @@ import { buildAdapterSettings } from '../../src/config'
 import { SubscriptionSetFactory } from '../../src/util'
 import { ExpiringSortedSet } from '../../src/util/subscription-set/expiring-sorted-set'
 import { RedisSubscriptionSet } from '../../src/util/subscription-set/redis-sorted-set'
-import { RedisMock } from '../util'
+import { RedisMock } from '../../src/util/testing-utils'
 
 test('subscription set factory (local cache)', async (t) => {
   process.env['CACHE_TYPE'] = 'local'
