+++ test/transports/routing.test.ts
@@ -17,7 +17,7 @@ import {
   WebSocketTransport,
 } from '../../src/transports'
 import { InputParameters } from '../../src/validation'
-import { TestAdapter, mockWebSocketProvider } from '../util'
+import { TestAdapter, mockWebSocketProvider } from '../../src/util/testing-utils'
 
 const test = untypedTest as TestFn<{
   testAdapter: TestAdapter<SettingsDefinitionFromConfig<typeof adapterConfig>>
