+++ test/transports/websocket.test.ts
@@ -11,7 +11,7 @@ import {
 } from '../../src/transports'
 import { SingleNumberResultResponse } from '../../src/util'
 import { InputParameters } from '../../src/validation'
-import { TestAdapter, mockWebSocketProvider, runAllUntilTime } from '../util'
+import { TestAdapter, mockWebSocketProvider, runAllUntilTime } from '../../src/util/testing-utils'
 
 export const test = untypedTest as TestFn<{
   testAdapter: TestAdapter
